feat: Separate api module
