+++ api/settings.gradle
@@ -1,0 +1,9 @@
+rootProject.name = 'api'
+
+dependencyResolutionManagement {
+    versionCatalogs {
+        libs {
+            from(files("../gradle/libs.versions.toml"))
+        }
+    }
+}
